Fix js_util compile errors via conditional web import
